feat(Bulk Edit): componerize
